% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodePeaks.R
\name{decodePeaks}
\alias{decodePeaks}
\title{Decode peaks}
\usage{
decodePeaks(x, compression = "none", size)
}
\arguments{
\item{x}{a \code{base64} encoded vector}

\item{compression}{compression type (default = "none")}

\item{size}{a numeric value for the number of bytes per element in the byte stream}
}
\value{
a numeric vector
}
\description{
Decode a \code{binaryDataArray} into a vector of either time or intensity
}
\keyword{internal}
