% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idRefs.R
\name{idRefs}
\alias{idRefs}
\title{Extract mzML idRefs}
\usage{
idRefs(xmlDoc)
}
\arguments{
\item{xmlDoc}{a `xml` document}
}
\value{
a character vector of all available `idRefs`
}
\description{
Extract all the `idRefs` in the specified `mzML` file. The `idRefs` will correspond to the individual scan events.
}
\examples{
\dontrun{
library(xml2)
xmlDoc <- read_xml("example_file.mzML")
xmlRefs <- idRefs(xmlDoc)
head(xmlRefs)
[1] "TIC"                     "SRM SIC 153.01,65.271"
[3] "SRM SIC 153.01,67.232"   "SRM SIC 153.01,109.094"
[5] "SRM SIC 179.022,107.007" "SRM SIC 179.022,134.006"
 }
}
