% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFile.R
\name{openFile}
\alias{openFile}
\title{Open .mzML File}
\usage{
openFile(mzml_file)
}
\arguments{
\item{mzml_file}{the absolute file path of a valid `.mzML` file}
}
\value{
a list of two elements containing `peaks` and `header`
}
\description{
Open and parse a .mzML file containing Selective Reaction Monitoring (SRM) Mass Spectrometry (MS) data using `xml2`
}
