% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChromNode.R
\name{parseChromNode}
\alias{parseChromNode}
\title{Parse the XML nodeset of a SRM chromatogram}
\usage{
parseChromNode(x, mode)
}
\arguments{
\item{x}{a `xml2` nodeset for a `Chromatogram` block}

\item{mode}{a character of either `SRM` or `TIC`}
}
\value{
a list of two elements
 * **time**
   * `attributes`
   * `raw`
 * **intensity**
   * `attributes`
   * `raw`
}
\description{
Parse the XML nodeset of a SRM chromatogram
}
