% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChromNode.R
\name{parseChromNode}
\alias{parseChromNode}
\title{Parse the XML nodeset of a SRM chromatogram}
\usage{
parseChromNode(x, mode)
}
\arguments{
\item{x}{a \code{xml2} nodeset for a \code{Chromatogram} block}

\item{mode}{a character of either \code{SRM} or \code{TIC}}
}
\value{
a list of two elements
\itemize{
\item \strong{time}
\itemize{
\item \code{attributes}
\item \code{raw}
}
\item \strong{intensity}
\itemize{
\item \code{attributes}
\item \code{raw}
}
}
}
\description{
Parse the XML nodeset of a SRM chromatogram
}
\keyword{internal}
